refactor: JS 로딩 방식 변경 및 파일 통합
