+++ public/community.html
@@ -742,7 +742,8 @@
       </li>
     </template>
 
-    <script src="./js/community.js" type="module"></script>
+    <!-- <script src="./js/community.js" type="module"></script> -->
+    <script src="./js/community.js"></script>
     <div data-include-path="./includes/footer.html"></div>
   </body>
 </html>
